Added controller for the overview page and the event handling

+++ NutritionTracker/NutritionTracker.Domain/DomainEntities/FoodEntity.cs
@@ -16,61 +16,61 @@ namespace NutritionTracker.Domain.DomainEntities
 		public string Group { get; set; }
 		public int? Energy_kcal { get; set; }
 		public int? Energy_kj { get; set; }
-		public float? FatTotal_g { get; set; }
-		public float? Protein_g { get; set; }
-		public float? Carbohydrates_g { get; set; }
-		public float? Fiber_g { get; set; }
-		public float? Water_g { get; set; }
-		public float? Alcohol_g { get; set; }
-		public float? Ash_g { get; set; }
-		public float? SugarsTotal_g { get; set; }
-		public float? Monosaccharides_g { get; set; }
-		public float? Disaccharides_g { get; set; }
+		public double? FatTotal_g { get; set; }
+		public double? Protein_g { get; set; }
+		public double? Carbohydrates_g { get; set; }
+		public double? Fiber_g { get; set; }
+		public double? Water_g { get; set; }
+		public double? Alcohol_g { get; set; }
+		public double? Ash_g { get; set; }
+		public double? SugarsTotal_g { get; set; }
+		public double? Monosaccharides_g { get; set; }
+		public double? Disaccharides_g { get; set; }
 		public int? AddedSugar_g { get; set; }
 		public int? FreeSugar_g { get; set; }
 		public int? WholeGrainTotal_g { get; set; }
-		public float? SaturatedFattyAcids_g { get; set; }
-		public float? FattyAcids410_g { get; set; }
-		public float? LauricAcid_g { get; set; }
-		public float? MyristicAcid_g { get; set; }
-		public float? PalmiticAcid_g { get; set; }
-		public float? StearicAcid_g { get; set; }
-		public float? ArachidicAcid_g { get; set; }
-		public float? MonounsaturatedFattyAcids_g { get; set; }
-		public float? PalmitoleicAcid_g { get; set; }
-		public float? OleicAcid_g { get; set; }
-		public float? PolyunsaturatedFattyAcids_g { get; set; }
-		public float? LinoleicAcid_g { get; set; }
-		public float? LinolenicAcid_g { get; set; }
-		public float? ArachidonicAcid_g { get; set; }
-		public float? EPA_g { get; set; }
-		public float? DPA_g { get; set; }
-		public float? DHA_g { get; set; }
-		public float? Cholesterol_mg { get; set; }
-		public float? Vitamin_A_Re_ug { get; set; }
-		public float? Retinol_ug { get; set; }
-		public float? BetaCarotene_ug { get; set; }
-		public float? Vitamin_D_ug { get; set; }
-		public float? Vitamin_E_mg { get; set; }
-		public float? Vitamin_K_ug { get; set; }
-		public float? Thiamin_mg { get; set; }
-		public float? Riboflavin_mg { get; set; }
-		public float? Niacin_mg { get; set; }
-		public float? NiacinEquivalents_mg { get; set; }
-		public float? Vitamin_B6_mg { get; set; }
-		public float? FolateTotal_ug { get; set; }
-		public float? Vitamin_B12_ug { get; set; }
-		public float? Vitamin_C_mg { get; set; }
+		public double? SaturatedFattyAcids_g { get; set; }
+		public double? FattyAcids410_g { get; set; }
+		public double? LauricAcid_g { get; set; }
+		public double? MyristicAcid_g { get; set; }
+		public double? PalmiticAcid_g { get; set; }
+		public double? StearicAcid_g { get; set; }
+		public double? ArachidicAcid_g { get; set; }
+		public double? MonounsaturatedFattyAcids_g { get; set; }
+		public double? PalmitoleicAcid_g { get; set; }
+		public double? OleicAcid_g { get; set; }
+		public double? PolyunsaturatedFattyAcids_g { get; set; }
+		public double? LinoleicAcid_g { get; set; }
+		public double? LinolenicAcid_g { get; set; }
+		public double? ArachidonicAcid_g { get; set; }
+		public double? EPA_g { get; set; }
+		public double? DPA_g { get; set; }
+		public double? DHA_g { get; set; }
+		public double? Cholesterol_mg { get; set; }
+		public double? Vitamin_A_Re_ug { get; set; }
+		public double? Retinol_ug { get; set; }
+		public double? BetaCarotene_ug { get; set; }
+		public double? Vitamin_D_ug { get; set; }
+		public double? Vitamin_E_mg { get; set; }
+		public double? Vitamin_K_ug { get; set; }
+		public double? Thiamin_mg { get; set; }
+		public double? Riboflavin_mg { get; set; }
+		public double? Niacin_mg { get; set; }
+		public double? NiacinEquivalents_mg { get; set; }
+		public double? Vitamin_B6_mg { get; set; }
+		public double? FolateTotal_ug { get; set; }
+		public double? Vitamin_B12_ug { get; set; }
+		public double? Vitamin_C_mg { get; set; }
 		public int? Phosphorus_mg { get; set; }
-		public float? Iodine_ug { get; set; }
-		public float? Iron_mg { get; set; }
+		public double? Iodine_ug { get; set; }
+		public double? Iron_mg { get; set; }
 		public int? Calcium_mg { get; set; }
 		public int? Potassium_mg { get; set; }
 		public int? Magnesium_mg { get; set; }
 		public int? Sodium_mg { get; set; }
-		public float? Salt_g { get; set; }
-		public float? Selenium_ug { get; set; }
-		public float? Zinc_mg { get; set; }
-		public float? Waste_percent { get; set; }
+		public double? Salt_g { get; set; }
+		public double? Selenium_ug { get; set; }
+		public double? Zinc_mg { get; set; }
+		public double? Waste_percent { get; set; }
 	}
 }
